api: add support for validating all cards
In case the processor used doesn't support $0 auth, trigger
a real auth (voided on success) to validate the card.

Pass the plugin property force_validation_amount=true to the auth
call to enable this behavior. The amount used is $1 by default but
can be changed using the force_validation_amount property.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>