api: return the actual amount used when verifying cards
Overriding the value only in memory is confusing since it doesn't
match the data on disk and could cause further confusion if the
payment is fixed by the Janitor (which would see the actual value).

Besides, this information might be interesting for the caller.

Related: https://github.com/killbill/killbill/issues/526.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>